movies/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import MovieViewSet, AIInsightViewSet

router = DefaultRouter()
router.register(r'movies', MovieViewSet)
router.register(r'ai-insights', AIInsightViewSet, basename='ai-insight')

urlpatterns = [
    path('', include(router.urls)),
]
