requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
beautifulsoup4==4.14.3
billiard==4.2.4
bs4==0.0.2
celery==5.3.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==47.0.0
distro==1.9.0
Django==6.0.4
django-filter==25.2
djangorestframework==3.17.1
drf-spectacular==0.28.0
google-auth==2.49.2
google-genai==1.73.1
groq==1.2.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
inflection==0.5.1
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
lxml==6.1.0
Markdown==3.10.2
openai==2.32.0
outcome==1.3.0.post0
packaging==26.2
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
redis==5.2.1
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
selenium==4.43.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
sqlparse==0.5.5
tenacity==9.1.4
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==14.2
whitenoise==6.9.0
wsproto==1.3.2
